Split AsyncSystem.h into multiple files.
