+++ CesiumAsync/CMakeLists.txt
@@ -55,3 +55,8 @@ install(TARGETS CesiumAsync
     LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
     PUBLIC_HEADER DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/CesiumAsync
 )
+
+install(
+    DIRECTORY include/CesiumAsync/Impl
+    DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/CesiumAsync
+)
