+++ CesiumAsync/include/CesiumAsync/AsyncSystem.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include "CesiumAsync/Future.h"
+#include "CesiumAsync/Impl/RemoveFuture.h"
+#include "CesiumAsync/Impl/WithTracing.h"
+#include "CesiumAsync/Impl/cesium-async++.h"
 #include "CesiumAsync/Library.h"
 #include "CesiumUtility/Profiler.h"
 #include <memory>
@@ -7,458 +11,10 @@
 #include <variant>
 #include <vector>
 
-#pragma warning(push)
-
-#pragma warning(disable : 4458 4324)
-
-#ifndef _MSC_VER
-#pragma GCC diagnostic ignored "-Wshadow"
-#endif
-
-#include <async++.h>
-
-#pragma warning(pop)
-
 namespace CesiumAsync {
 class ITaskProcessor;
 
-template <class T> class Future;
-
-namespace Impl {
-// Begin omitting doxgen warnings for Impl namespace
-//! @cond Doxygen_Suppress
-
-class AsyncSystemSchedulers;
-
-class ImmediatelyInMainThreadScheduler {
-public:
-  ImmediatelyInMainThreadScheduler(AsyncSystemSchedulers* pSchedulers)
-      : _pSchedulers(pSchedulers) {}
-
-  void schedule(async::task_run_handle t);
-
-private:
-  AsyncSystemSchedulers* _pSchedulers;
-};
-
-class WorkerThreadScheduler {
-public:
-  WorkerThreadScheduler(AsyncSystemSchedulers* pSchedulers)
-      : _pSchedulers(pSchedulers) {}
-
-  void schedule(async::task_run_handle t);
-
-private:
-  AsyncSystemSchedulers* _pSchedulers;
-};
-
-class ImmediatelyInWorkerThreadScheduler {
-public:
-  ImmediatelyInWorkerThreadScheduler(AsyncSystemSchedulers* pSchedulers)
-      : _pSchedulers(pSchedulers) {}
-
-  void schedule(async::task_run_handle t);
-
-private:
-  AsyncSystemSchedulers* _pSchedulers;
-};
-
-class AsyncSystemSchedulers {
-public:
-  AsyncSystemSchedulers(const std::shared_ptr<ITaskProcessor>& pTaskProcessor);
-
-  std::shared_ptr<ITaskProcessor> pTaskProcessor;
-
-  async::fifo_scheduler mainThreadScheduler;
-  WorkerThreadScheduler workerThreadScheduler;
-
-  ImmediatelyInMainThreadScheduler immediatelyInMainThreadScheduler;
-  ImmediatelyInWorkerThreadScheduler immediatelyInWorkerThreadScheduler;
-
-  void startInMainThread();
-  void endInMainThread();
-
-  void startInWorkerThread();
-  void endInWorkerThread();
-
-  static thread_local std::vector<AsyncSystemSchedulers*>
-      schedulersInMainThread;
-  static thread_local std::vector<AsyncSystemSchedulers*>
-      schedulersInWorkerThread;
-};
-
-template <typename T> struct RemoveFuture { typedef T type; };
-template <typename T> struct RemoveFuture<Future<T>> { typedef T type; };
-template <typename T> struct RemoveFuture<const Future<T>> { typedef T type; };
-template <typename T> struct RemoveFuture<async::task<T>> { typedef T type; };
-template <typename T> struct RemoveFuture<const async::task<T>> {
-  typedef T type;
-};
-
-template <class Func> struct IdentityUnwrapper {
-  static Func unwrap(Func&& f) { return std::forward<Func>(f); }
-};
-
-template <class Func, class T> struct ParameterizedTaskUnwrapper {
-  static auto unwrap(Func&& f) {
-    return [f = std::move(f)](T&& t) { return f(std::forward<T>(t))._task; };
-  }
-};
-
-template <class Func> struct TaskUnwrapper {
-  static auto unwrap(Func&& f) {
-    return [f = std::move(f)]() { return f()._task; };
-  }
-};
-
-template <typename Func, typename T> struct ContinuationReturnType {
-  typedef typename std::invoke_result<Func, T>::type type;
-};
-
-template <typename Func> struct ContinuationReturnType<Func, void> {
-  typedef typename std::invoke_result<Func>::type type;
-};
-
-template <typename Func, typename T> struct ContinuationFutureType {
-  typedef typename RemoveFuture<
-      typename ContinuationReturnType<Func, T>::type>::type type;
-};
-
-template <class Func, class T> auto unwrapFuture(Func&& f) {
-  return std::conditional<
-      std::is_same<
-          typename ContinuationReturnType<Func, T>::type,
-          typename ContinuationFutureType<Func, T>::type>::value,
-      IdentityUnwrapper<Func>,
-      ParameterizedTaskUnwrapper<Func, T>>::type::unwrap(std::forward<Func>(f));
-}
-
-template <class Func> auto unwrapFuture(Func&& f) {
-  return std::conditional<
-      std::is_same<
-          typename ContinuationReturnType<Func, void>::type,
-          typename ContinuationFutureType<Func, void>::type>::value,
-      IdentityUnwrapper<Func>,
-      TaskUnwrapper<Func>>::type::unwrap(std::forward<Func>(f));
-}
-
-template <typename Func, typename T> struct WithTracing {
-  static auto wrap(const char* tracingName, Func&& f) {
-#if TRACING_ENABLED
-    int64_t tracingID = CesiumUtility::Profiler::instance().getEnlistedID();
-
-    return [tracingID,
-            tracingName,
-            f = Impl::unwrapFuture<Func, T>(std::forward<Func>(f))](
-               T&& result) mutable {
-      TRACE_ASYNC_ENLIST(tracingID);
-      if (tracingName) {
-        TRACE_ASYNC_END_ID(tracingName, tracingID);
-      }
-      return f(std::move(result));
-    };
-#else
-    return Impl::unwrapFuture<Func, T>(std::forward<Func>(f));
-#endif
-  }
-};
-
-template <typename Func> struct WithTracing<Func, void> {
-  static auto wrap(const char* tracingName, Func&& f) {
-#if TRACING_ENABLED
-    int64_t tracingID = CesiumUtility::Profiler::instance().getEnlistedID();
-
-    return [tracingID,
-            tracingName,
-            f = Impl::unwrapFuture<Func>(std::forward<Func>(f))]() mutable {
-      TRACE_ASYNC_ENLIST(tracingID);
-      if (tracingName) {
-        TRACE_ASYNC_END_ID(tracingName, tracingID);
-      }
-      return f();
-    };
-#else
-    return Impl::unwrapFuture<Func, T>(std::forward<Func>(f));
-#endif
-  }
-};
-
-template <typename Func, typename T, typename Scheduler> struct CatchFunction {
-  Scheduler& scheduler;
-  Func f;
-  std::shared_ptr<ITaskProcessor> pTaskProcessor;
-
-  decltype(auto) operator()(async::task<T>&& t) {
-    try {
-      return async::make_task(t.get());
-    } catch (std::exception& e) {
-      return async::make_task(e).then(
-          scheduler,
-          unwrapFuture<Func, std::exception>(std::move(f)));
-    } catch (...) {
-      return async::make_task<std::exception>(
-                 std::runtime_error("Unknown exception"))
-          .then(scheduler, unwrapFuture<Func, std::exception>(std::move(f)));
-    }
-  }
-};
-
-template <typename Func, typename Scheduler>
-struct CatchFunction<Func, void, Scheduler> {
-  Scheduler& scheduler;
-  Func f;
-  std::shared_ptr<ITaskProcessor> pTaskProcessor;
-
-  decltype(auto) operator()(async::task<void>&& t) {
-    try {
-      t.get();
-      return async::make_task();
-    } catch (std::exception& e) {
-      return async::make_task(e).then(
-          scheduler,
-          unwrapFuture<Func, std::exception>(std::move(f)));
-    } catch (...) {
-      return async::make_task<std::exception>(
-                 std::runtime_error("Unknown exception"))
-          .then(scheduler, unwrapFuture<Func, std::exception>(std::move(f)));
-    }
-  }
-};
-
-//! @endcond
-// End omitting doxgen warnings for Impl namespace
-} // namespace Impl
-
-/**
- * @brief A value that will be available in the future, as produced by
- * {@link AsyncSystem}.
- *
- * @tparam T The type of the value.
- */
-template <class T> class Future final {
-public:
-  /**
-   * @brief Move constructor
-   */
-  Future(Future<T>&& rhs) noexcept
-      : _pSchedulers(std::move(rhs._pSchedulers)),
-        _task(std::move(rhs._task)) {}
-
-  Future(const Future<T>& rhs) = delete;
-  Future<T>& operator=(const Future<T>& rhs) = delete;
-
-  /**
-   * @brief Registers a continuation function to be invoked in a worker thread
-   * when this Future resolves, and invalidates this Future.
-   *
-   * If the function itself returns a `Future`, the function will not be
-   * considered complete until that returned `Future` also resolves.
-   *
-   * Even if the previous Future resolves
-   *
-   * @tparam Func The type of the function.
-   * @param f The function.
-   * @return A future that resolves after the supplied function completes.
-   */
-  template <class Func>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenInWorkerThread(Func&& f) && {
-    return std::move(*this).thenWithScheduler(
-        this->_pSchedulers->workerThreadScheduler,
-        "waiting for worker thread",
-        std::move(f));
-  }
-
-  /**
-   * @brief Registers a continuation function to be invoked in the main thread
-   * when this Future resolves, and invalidates this Future.
-   *
-   * The supplied function will not be invoked immediately, even if this method
-   * is called from the main thread. Instead, it will be queued and invoked the
-   * next time {@link AsyncSystem::dispatchMainThreadTasks} is called.
-   *
-   * If the function itself returns a `Future`, the function will not be
-   * considered complete until that returned `Future` also resolves.
-   *
-   * @tparam Func The type of the function.
-   * @param f The function.
-   * @return A future that resolves after the supplied function completes.
-   */
-  template <class Func>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenInMainThread(Func&& f) && {
-    return std::move(*this).thenWithScheduler(
-        this->_pSchedulers->mainThreadScheduler,
-        "waiting for main thread",
-        std::move(f));
-  }
-
-  template <class Func>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenImmediatelyInMainThread(Func&& f) && {
-    return std::move(*this).thenWithScheduler(
-        this->_pSchedulers->immediatelyInMainThreadScheduler,
-        "waiting for main thread",
-        std::move(f));
-  }
-
-  /**
-   * @brief Registers a continuation function to be invoked immediately in
-   * whichever thread causes the Future to be resolved, and invalidates this
-   * Future.
-   *
-   * If the Future is already resolved, the supplied function will be called
-   * immediately in the calling thread and this method will not return until
-   * that function does.
-   *
-   * If the function itself returns a `Future`, the function will not be
-   * considered complete until that returned `Future` also resolves.
-   *
-   * @tparam Func The type of the function.
-   * @param f The function.
-   * @return A future that resolves after the supplied function completes.
-   */
-  template <class Func>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenImmediately(Func&& f) && {
-    return Future<typename Impl::ContinuationFutureType<Func, T>::type>(
-        this->_pSchedulers,
-        _task.then(
-            async::inline_scheduler(),
-            Impl::WithTracing<Func, T>::wrap(nullptr, std::forward<Func>(f))));
-  }
-
-  /**
-   * @brief Registers a continuation function to be invoked in a worker thread
-   * when this Future resolves, and invalidates this Future.
-   *
-   * If the function itself returns a `Future`, the function will not be
-   * considered complete until that returned `Future` also resolves.
-   *
-   * Unlike {@link thenInWorkerThread},
-   *
-   * @tparam Func The type of the function.
-   * @param f The function.
-   * @return A future that resolves after the supplied function completes.
-   */
-  template <class Func>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenImmediatelyInWorkerThread(Func&& f) && {
-    return std::move(*this).thenWithScheduler(
-        this->_pSchedulers->immediatelyInWorkerThreadScheduler,
-        "waiting for worker thread",
-        std::move(f));
-  }
-
-  /**
-   * @brief Registers a continuation function to be invoked in the main thread
-   * when this Future rejects, and invalidates this Future.
-   *
-   * The supplied function will not be invoked immediately, even if this method
-   * is called from the main thread. Instead, it will be queued and invoked the
-   * next time {@link AsyncSystem::dispatchMainThreadTasks} is called.
-   *
-   * If the function itself returns a `Future`, the function will not be
-   * considered complete until that returned `Future` also resolves.
-   *
-   * Any `then` continuations chained after this one will be invoked with the
-   * return value of the catch callback.
-   *
-   * @tparam Func The type of the function.
-   * @param f The function.
-   * @return A future that resolves after the supplied function completes.
-   */
-  template <class Func> Future<T> catchInMainThread(Func&& f) && {
-    return std::move(*this).catchWithScheduler(
-        this->_pSchedulers->mainThreadScheduler,
-        std::move(f));
-  }
-
-  template <class Func, class Scheduler>
-  Future<typename Impl::ContinuationFutureType<Func, std::exception>::type>
-  catchWithScheduler(Scheduler& scheduler, Func&& f) && {
-    return Future<
-        typename Impl::ContinuationFutureType<Func, std::exception>::type>(
-        this->_pSchedulers,
-        this->_task.then(
-            async::inline_scheduler(),
-            Impl::CatchFunction<Func, T, Scheduler>{
-                scheduler,
-                std::move(f),
-                this->_pSchedulers->pTaskProcessor}));
-  }
-
-  /**
-   * @brief Waits for the future to resolve or reject and returns the result.
-   *
-   * This method must not be called from the main thread, the one that calls
-   * {@link AsyncSystem::dispatchMainThreadTasks}. Doing so can lead to a
-   * deadlock because the main thread tasks will never complete while this
-   * method is blocking the main thread.
-   *
-   * @return The value if the future resolves successfully, or the exception if
-   * it rejects.
-   */
-  std::variant<T, std::exception> wait() {
-    try {
-      return this->_task.get();
-    } catch (std::exception& e) {
-      return e;
-    } catch (...) {
-      return std::runtime_error("Unknown exception.");
-    }
-  }
-
-private:
-  Future(
-      const std::shared_ptr<Impl::AsyncSystemSchedulers>& pSchedulers,
-      async::task<T>&& task) noexcept
-      : _pSchedulers(pSchedulers), _task(std::move(task)) {}
-
-  template <class Func, class Scheduler>
-  Future<typename Impl::ContinuationFutureType<Func, T>::type>
-  thenWithScheduler(
-      Scheduler& scheduler,
-      const char* tracingName,
-      Func&& f) && {
-    // It would be nice if tracingName were a template parameter instead of a
-    // function parameter, but that triggers a bug in VS2017. It was previously
-    // a bug in VS2019, too, but has been fixed there:
-    // https://developercommunity.visualstudio.com/t/internal-compiler-error-when-compiling-a-template-1/534210
-#if TRACING_ENABLED
-    // When tracing is enabled, we measure the time between scheduling and
-    // dispatching of the work.
-    int64_t tracingID = CesiumUtility::Profiler::instance().getEnlistedID();
-    auto task = this->_task.then(
-        async::inline_scheduler(),
-        [tracingID, tracingName](T&& value) {
-          TRACE_ASYNC_BEGIN_ID(tracingName, tracingID);
-          return std::move(value);
-        });
-#else
-    auto& task = this->_task;
-#endif
-
-    return Future<typename Impl::ContinuationFutureType<Func, T>::type>(
-        this->_pSchedulers,
-        task.then(
-            scheduler,
-            Impl::WithTracing<Func, T>::wrap(
-                tracingName,
-                std::forward<Func>(f))));
-  }
-
-  std::shared_ptr<Impl::AsyncSystemSchedulers> _pSchedulers;
-  async::task<T> _task;
-
-  friend class AsyncSystem;
-
-  template <class Func, class R> friend struct Impl::ParameterizedTaskUnwrapper;
-
-  template <class Func> friend struct Impl::TaskUnwrapper;
-
-  template <class R> friend class Future;
-};
+template <typename T> class Future;
 
 /**
  * @brief A system for managing asynchronous requests and tasks.
@@ -505,7 +61,7 @@ public:
    */
   AsyncSystem(const std::shared_ptr<ITaskProcessor>& pTaskProcessor) noexcept;
 
-  template <class T, class Func> Future<T> createFuture(Func&& f) const {
+  template <typename T, typename Func> Future<T> createFuture(Func&& f) const {
     std::shared_ptr<async::event_task<T>> pEvent =
         std::make_shared<async::event_task<T>>();
 
@@ -588,7 +144,7 @@ public:
    * @param value The value for the future
    * @return The future
    */
-  template <class T> Future<T> createResolvedFuture(T&& value) const {
+  template <typename T> Future<T> createResolvedFuture(T&& value) const {
     return Future<T>(
         this->_pSchedulers,
         async::make_task<T>(std::forward<T>(value)));
@@ -608,6 +164,6 @@ public:
 private:
   std::shared_ptr<Impl::AsyncSystemSchedulers> _pSchedulers;
 
-  template <class T> friend class Future;
+  template <typename T> friend class Future;
 };
 } // namespace CesiumAsync
