+++ CesiumAsync/include/CesiumAsync/Impl/cesium-async++.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#pragma warning(push)
+
+#pragma warning(disable : 4458 4324)
+
+#ifndef _MSC_VER
+#pragma GCC diagnostic ignored "-Wshadow"
+#endif
+
+#include <async++.h>
+
+#pragma warning(pop)
